clear old unneeded CRUD classes #81
