--- classes/testing/crud/price_test.php
@@ -1,239 +1,0 @@
-<?php
-/**
- * Price test CRUD
- *
- * @package   ingot
- * @author    Josh Pollock <Josh@JoshPress.net>
- * @license   GPL-2.0+
- * @link
- * @copyright 2015 Josh Pollock
- */
-namespace ingot\testing\crud;
-
-
-use ingot\testing\utility\defaults;
-use ingot\testing\utility\price;
-
-class price_test extends options_crud {
-
-	/**
-	 * Name of this object
-	 *
-	 * @since 0.0.9
-	 *
-	 * @access protected
-	 *
-	 * @var string
-	 */
-	protected static $what = 'price_test_group';
-
-	protected static function what() {
-		return self::$what;
-	}
-
-	/**
-	 * Get price tests
-	 *
-	 * @since 0.0.9
-	 *
-	 * @acess protected
-	 *
-	 * @param array $params {
-	 *  $ids bool Optional. If false, the default, all fields are returned. If true, only IDs
-	 *  $limit int Optional. Total number to return. Default is -1 to get all.
-	 *  $page int Optional. Page of results. Default is 1
-	 * }
-	 *
-	 * @return array|null|object
-	 */
-	public static function get_items( $params ) {
-		$args = wp_parse_args(
-			$params,
-			array(
-				'ids' => false,
-				'limit' => -1,
-				'page' => 1
-			)
-		);
-
-		if( ! empty(  $args[ 'ids' ] ) && is_array( $args[ 'ids' ] ) ) {
-			global $wpdb;
-
-			$what = self::$what;
-			$pattern = '`option_name` = "ingot_%s_%s"';
-
-			foreach( $args[ 'ids' ] as $id ){
-				$or[]= sprintf( $pattern, $what, $id );
-			}
-
-			$or = implode( ' OR', $or );
-
-			$sql = sprintf( "SELECT * FROM %s WHERE %s", $wpdb->options, $or );
-
-			$results = $wpdb->get_results( $sql, ARRAY_A );
-
-			if ( ! empty( $results ) ) {
-				$tests = array();
-				foreach( $results as $result ){
-					$tests[] = self::fill_in( unserialize( $result[ 'option_value' ] ) );
-				}
-
-				return $tests;
-
-			}
-
-		}
-
-	}
-
-
-	/**
-	 *  Prepared data to be saved
-	 *
-	 * @since 0.0.9
-	 *
-	 * @access protected
-	 *
-	 * @param array $data Data
-	 *
-	 * @return array|\WP_Error Array of prepared data or WP_Error if invalid.
-	 */
-	protected static function validate_config( $data ) {
-		if( ! isset( $data[ 'default' ] ) ) {
-			return new \WP_Error( 'ingot-invalid-price-test', __( 'Price tests must have a default set of price variations' ), 'ingot' );
-		}
-
-		if( ! price::valid_percentage( $data[ 'default'] ) ) {
-			return new \WP_Error( 'ingot-invalid-float' );
-
-		}
-
-		$data = self::fill_in( $data );
-
-		if( is_array( $data[ 'variable_prices' ] ) && ! empty( $data[ 'variable_prices' ] ) ){
-			foreach( $data[ 'variable_prices' ] as $i => $test_part ) {
-
-				if( ! price::valid_percentage( $test_part ) ) {
-					return new \WP_Error( 'ingot-invalid-float' );
-				}
-
-				$data[ 'variable_prices' ][ $i ] = $test_part;
-			}
-		}
-
-		if( ! isset( $data[ 'created' ] ) || 0 == $data[ 'created' ] ) {
-			$data[ 'created' ] = time();
-
-		}
-
-		if( ! isset( $data[ 'modified' ] ) || 0 == $data[ 'modified' ] ) {
-			$data[ 'modified' ] = time();
-		}
-
-
-
-		return $data;
-
-
-	}
-
-
-
-
-	/**
-	 * Fill in needed, but not required keys
-	 * @since 0.0.4
-	 *
-	 * @access protected
-	 *
-	 * @param $data
-	 *
-	 * @return array
-	 */
-	protected static function fill_in( $data ){
-		if( empty( $data[ 'threshold' ] ) ) {
-			$data[ 'threshold' ] = defaults::threshold();
-		}
-
-		if( empty( $data[ 'initial'] ) ){
-			$data[ 'initial' ] = defaults::initial();
-		}
-
-		foreach( self::needed() as $needed ) {
-			if( ! isset( $data[ $needed ] ) ) {
-				if( 'variable_prices' == $needed ) {
-					$data[ 'variable_prices' ] = array();
-				}else {
-					$data[ $needed ] = 0;
-				}
-			}
-
-		}
-
-
-
-		return $data;
-
-	}
-
-
-	/**
-	 * Check that we have a float less than 1 and greater than -1
-	 *
-	 * @param float $i
-	 *
-	 * @return bool
-	 */
-	protected static function float_in_allowed_range( $i ) {
-		if( $i > -1 && $i < 1 ) {
-			return true;
-
-		}else{
-			return false;
-
-		}
-
-	}
-
-
-	/**
-	 * Required fields of this object
-	 *
-	 * @since 0.0.9
-	 *
-	 * @access protected
-	 *
-	 * @return array
-	 */
-	protected static function required() {
-		$required = array(
-			'product_ID',
-			'default'
-		);
-
-		return $required;
-
-	}
-
-	/**
-	 * Neccasary, but not required fields of this object
-	 *
-	 * @since 0.0.9
-	 *
-	 * @access protected
-	 *
-	 * @return array
-	 */
-	protected static function needed() {
-
-		$needed = array(
-			'name',
-			'variable_prices',
-			'created',
-			'modified'
-		);
-
-		return $needed;
-
-	}
-}
